added reliability progress
